REF: Remove info.py and implement automatic versioning
