+++ nipype/pkg_info.py
@@ -8,27 +8,11 @@ COMMIT_INFO_FNAME = "COMMIT_INFO.txt"
 
 
 def pkg_commit_hash(pkg_path):
-    """Get short form of commit hash given directory `pkg_path`
-
-    There should be a file called 'COMMIT_INFO.txt' in `pkg_path`.  This is a
-    file in INI file format, with at least one section: ``commit hash`` and two
-    variables ``archive_subst_hash`` and ``install_hash``.  The first has a
-    substitution pattern in it which may have been filled by the execution of
-    ``git archive`` if this is an archive generated that way.  The second is
-    filled in by the installation, if the installation is from a git archive.
+    """Get short form of commit hash.
 
     We get the commit hash from (in order of preference):
-
-    * A substituted value in ``archive_subst_hash``
-    * A written commit hash value in ``install_hash`
+    * The local part of the version string (if installed)
     * git's output, if we are in a git repository
-
-    If all these fail, we return a not-found placeholder tuple
-
-    Parameters
-    ----------
-    pkg_path : str
-       directory containing package
 
     Returns
     -------
@@ -37,30 +21,31 @@ def pkg_commit_hash(pkg_path):
     hash_str : str
        short form of hash
     """
-    # Try and get commit from written commit text file
-    pth = os.path.join(pkg_path, COMMIT_INFO_FNAME)
-    if not os.path.isfile(pth):
-        raise OSError("Missing commit info file %s" % pth)
-    cfg_parser = configparser.RawConfigParser()
-    with open(pth, encoding="utf-8") as fp:
-        cfg_parser.read_file(fp)
-    archive_subst = cfg_parser.get("commit hash", "archive_subst_hash")
-    if not archive_subst.startswith("$Format"):  # it has been substituted
-        return "archive substitution", archive_subst
-    install_subst = cfg_parser.get("commit hash", "install_hash")
-    if install_subst != "":
-        return "installation", install_subst
+    from . import __version__
+    
+    # If version has a local part (e.g. +g8234ec318)
+    if "+" in __version__:
+        local_part = __version__.split("+")[1]
+        # hatch-vcs/setuptools-scm format: g<hash>
+        if local_part.startswith("g"):
+            hsh = local_part[1:].split(".")[0]
+            return "installation", hsh
+
     # maybe we are in a repository
-    proc = subprocess.Popen(
-        "git rev-parse --short HEAD",
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        cwd=pkg_path,
-        shell=True,
-    )
-    repo_commit, _ = proc.communicate()
-    if repo_commit:
-        return "repository", repo_commit.decode().strip()
+    try:
+        proc = subprocess.Popen(
+            "git rev-parse --short HEAD",
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            cwd=pkg_path,
+            shell=True,
+        )
+        repo_commit, _ = proc.communicate()
+        if repo_commit:
+            return "repository", repo_commit.decode().strip()
+    except Exception:
+        pass
+
     return "(none found)", "<not found>"
 
 
@@ -78,7 +63,7 @@ def get_pkg_info(pkg_path):
        with named parameters of interest
     """
     src, hsh = pkg_commit_hash(pkg_path)
-    from .info import VERSION
+    from . import __version__ as VERSION
     import networkx
     import nibabel
     import numpy
